2009-07-15  Sebastien Pouliot  <sebastien@ximian.com>
	* gdipFunctions.cs: Fix XVisualIDFromVisual to return an IntPtr
	since 64 bits X servers will return a 64 bits value.
	* gdipStructs.cs: Fix XVisualInfo to use IntPtr on fields that
	are 64 bits on 64 bits X servers.
	* Graphics.cs: Adjust CopyFromScreenX11 to work with IntPtr so the
	screenshot code can work on both 32 and 64 bits X servers.
	[Fix bug #521718]
	[Backport r137982]


svn path=/branches/mono-2-4/mcs/; revision=137983

+++ mcs/class/System.Drawing/System.Drawing/gdipStructs.cs
@@ -198,13 +198,13 @@ namespace System.Drawing
 	internal struct XVisualInfo
 	{
 		internal IntPtr visual;
-		internal int visualid;
+		internal IntPtr visualid;
 		internal int screen;
 		internal uint depth;
 		internal int klass;
-		internal uint red_mask;
-		internal uint green_mask;
-		internal uint blue_mask;
+		internal IntPtr red_mask;
+		internal IntPtr green_mask;
+		internal IntPtr blue_mask;
 		internal int colormap_size;
 		internal int bits_per_rgb;		
 	}
